Copied file from computer to github
This is an upsupervised classification that is automated to run on a batch of files from a particular location. This is run in TerrSet and the values may be changed depending on the goals of your personal project. This also includes a line for an NDVI mask to first block out and state which things are vegetation according to a particular NDVI range.